Add translation of array types.
For the time being, the translation discards any constraints
imposed on the compoment of the array during its declaration.

The current app also leaks memory, but the translation results
will be refcounted in the future to avoid the need for explicit
memory managment from the user of the lib.

no-tn-check

Change-Id: I235a6db2bfc4b2aeb28d5e1dce5a7692d26f2dd1